feat: Implemented log trace.
Currently, there is only an example of how to send notifications, which can be extended to other methods depending on the chosen approach.

Additionally, tests were added to test if the notification is not sent when tracing is off, only the message is sent when set to Messages, and verbosity is added when set to Verbose.